+++ src/core/api/users/users.api.ts
@@ -1,4 +1,4 @@
-import { UserReq, User, StripeLinkRes, StripeProfileRes, ForgetPasswordReq, StripeLinkReq } from './users.types';
+import { UserReq, User, ForgetPasswordReq } from './users.types';
 import { get, put } from '../http';
 import { SuccessRes } from '../types';
 
@@ -12,12 +12,4 @@ export async function updateProfile(payload: UserReq): Promise<User> {
 
 export async function updatePassword(payload: ForgetPasswordReq): Promise<SuccessRes> {
   return (await put<SuccessRes>('auth/password', payload)).data;
-}
-
-export async function getStripeLink(params: StripeLinkReq): Promise<StripeLinkRes> {
-  return (await get<StripeLinkRes>('auth/stripe/connect-link', { params })).data;
-}
-
-export async function getUserStripeProfile(): Promise<StripeProfileRes> {
-  return (await get<StripeProfileRes>('auth/stripe/profile')).data;
 }
