+++ src/core/api/users/users.types.ts
@@ -21,33 +21,3 @@ export interface ForgetPasswordReq {
   current_password?: string;
   password: string;
 }
-
-export interface StripeLinkReq {
-  country: string;
-  redirect_url?: string;
-}
-
-export interface StripeLinkRes {
-  link: {
-    url: string;
-    expires_at: Date;
-  };
-}
-
-export interface StripeAccountRes {
-  id: string;
-  account: string;
-  account_holder_name: string;
-  account_holder_type: null;
-  account_type: string;
-  bank_name: string;
-  last4: string;
-}
-
-export interface StripeProfileRes {
-  id: string;
-  external_accounts: {
-    data: StripeAccountRes[];
-    url: string;
-  };
-}
